Time: 0 ms (100.00%), Space: 41.1 MB (80.12%) - LeetHub

+++ 1491-average-salary-excluding-the-minimum-and-maximum-salary/1491-average-salary-excluding-the-minimum-and-maximum-salary.java
@@ -12,6 +12,6 @@ class Solution {
         }
         
                
-        return (double)((sum-min-max)/(salary.length - 2));
+        return ((sum-min-max)/(salary.length - 2));
     }
 }
